Treat planar curves as a special case when generating 3D splines
Treating each dimension x,y,z as independent when generating a 3D
spline can produce poor behavior if the data is far from even.
Fixing this in the general case is hard, but for the special case of
a curve lying in the xy, xz, or xy plane we can do better by treating
it as a 2D spline with only one independent variable.
This case is relatively common, e.g. for fence plots.