create separate section of the manual for "show"
Most "show" options are complementary to a corresponding "set" command.
Documenting them together makes sense. However a few "show" commands do
not correspond directly to an obvious "set".  These can be hard to find
in the current help system.  Move them to a separate, indexed, section
of the manual.
In particular this expands and indexes the various "show palette" options.