change df_read_matrix to return an array of doubles, not floats
Currently all binary data is read in and stored as (float) rather
than (double), and storage of text matrix data conforms to this.
Unfortunately this means that a matrix of 32bit integers is corrupted
by being stored in a float with only 24 bits of precision.
Ideally I would like to change the data storage to (double) everywhere,
but for now change it only for text matrix data.