Private implementation of VP_fwhm
The libcerf version of this function is not robust and is unacceptably
prone to terminating the program by calling exit(-1) on non-convergence.
We can do much better by using an algorithm guaranteed to converge.
The convergence criteria is currently set to 2*DBL_EPSILON, which is
probably overkill.
Replaces commits 809893f3 and 4b83330a.