Make sure "exit" works from inside a bracketed clause
The commands in a bracketed clause are handled internally using the
same mechanism as the commands in a loaded file.  When "exit" was
encountered, this mechanism only popped the most recent clause rather
than the most recent file in which the clause occurred.
Now we make sure that "exit" forces an early termination of the
clause as well as popping any nested clauses until we reach an
actual loaded file or the top level of interactive input.

Bug #2239