prevent overflow of data read from an ascii matrix file into an image
If unexpected data is encountered in the middle of reading an ascii
matrix input stream, this could overflow the space allocated for an
image.  This patch adds an error check to detect overflow.
It does *not* fix the handling of extra data, it just prevents the
overflow and consequent possible segfault.