track blank lines in ascii matrix input file
Without this the program did not reliably terminate input of the first
datablock, meaning that garbage (or in the bug report columnheaders)
on the line following the 2nd blank was treated as part of the first
data set or a spurious new data set (not sure which but either way it
caused bad indexing and a segfault if passed to "with image").
Now we count blanks, terminate on the 2nd blank, and ignore any further blanks.

Note:  The program does not expect or handle columnheaders for any data
block except the first one. It will typically generate an warning
about bad data and store a spurious row of data that manifests as an
extra row of pixels if this is treated as image data.