add a variant of cspline smoothing for 2D curves that are not monotonic in x
    plot <data> using 1:2 smooth csplines with filledcurves closed
 or
    plot <data> using 1:2 smooth path with {filledcurves|lines}

"with filledcurves closed" will add a point if necessary to make the path
really closed.  That allows the spline fit to wrap around smoothly,
creating a closed curve that is smooth everywhere.

"with lines" will wrap around to make a closed shape if the last point
already matches the first point, but otherwise will create an open-ended
smooth curve rather than a closed one.  This allows loops, knots, etc.

TODO:
- The number of splines per closed curved should be controlled
  by something other than "set samples"