reset plot line properties after drawing errorbars
errorbars can have line properties (e.g. thin dashed line) distinct
from the principle line and point of the corresponding plot.
If these are not reset after drawing a key symbol or between drawing
the errorbars and the points in "with [x][y]errorlines" and
"with [x][y]errorbars" then the errorbar line properties incorrectly
replace those of the main plot.

Bug #2251