New function invigamma(a,p)
The inverse incomplete gamma function `invigamma(a,p)` returns the value
z such that p = igamma(a,z).

p is limited to (0;1]. a is limited to positive real values.

The implementation uses Halley's method to iterate starting from
approximations given by Abramowitz & Stegun.
The relative accuracy ranges from 1.e-16 for a<1 to 5.e-6 for a = 1.e10.
Convergence may fail for a < 0.005.