complex exponential integral expint(n,z)
The existing implementation of expint is limited to real z >= 0.
Now we add a configuration option --with-amos=<library_dir>
that searches for a support library containing the Amos routine cexint.
If found, gnuplot uses it to extend the domain of expint(n,z) to
complex z for n>0.

If a suitable library is not found during configuration, or if
n==0 at runtime, we fall back to the existing routine expint().

Reference: Amos 1990 Algorithm 683, ACM Trans Math Software 16:178].