Change handling of NaN/undefined points input to 3D plot modes
Since version 5 2D plot modes have stored input data into a point's
data structure even if one of the requested columns was NaN or UNDEFINED.
Now we extend this behavior to 3D plot modes as well.
This means, for example, that x and y can be valid even if z is
NaN or Inf or UNDEFINED.  Prior to this x and y contained junk in this case.

Motivation:  Allow interpolation across "holes" in a pm3d surface
caused by missing data.