special function SynchrotronF(x)
The synchrotron function F(x) describes the power spectrum of
synchrotron radiation as a function of photon energy (frequency)
in units of the critical frequency.  Chebyshev coefficients for
approximation of F(x) accurate to 1.E-16 are taken from
MacLeod (2000) NuclInstMethPhysRes A443.
The complementary synchrotron function G(x) can be calculated directly as
    G(x) = x * BesselK(2/3., x)  so no separate implementation is needed.