Fix memory leaks after unexpected input read by const_express()
The routine const_express() dates back to a time when the only possible
value types returned by an expression were integer and complex, so the
original callers did not check for or deal with anything else.
When newer types STRING ARRAY DATABLOCK etc were added this introduced
the possibility of a memory leak if the user provided a string or array
name where a numerical input value was expected.  Most such places were
changed to use wrappers int_expression() or real_expression().
This commit fixes all the remaining places in the core code and in
non-legacy terminals.

Motivation:  prevent memory leaks; prepare for introduction of functions
or operators that return an ARRAY value.