Allow evaluation stack to return an ARRAY value.
This is tricky because evaluation must delete any temporary array structures
used internally, but must not corrupt or delete existing arrays passed in
as parameters.  The new ARRAY subtype TEMP_ARRAY is used to flag arrays
created by the evaluation stack.  Whenever the stack dereferences an array
via pop() the flag is checked; a TEMP_ARRAY is freed after use but a
permanent array is left untouched.

Top level callers must be prepared to deal with const_express() returning
an ARRAY, either freeing it if it is not wanted or erasing the TEMP_ARRAY
flag if it is kept.

Note that prior to this it was only possible to create an array by
declaring it, perhaps with an initializer.  Now any variable will become
an array if set to an expression that evaluates to type ARRAY,
e.g. "C = A" in the example.  However array element types are still
limited to integer, complex, or string.