tikz: force implicit border width of text nodes to 0pt
The tikz/pgf justification of text nodes apparently allows space for a
border with current linewidth, even though no border is drawn.
Since gnuplot doesn't expect this, there is no guarantee what the
current linewidth is when it sends text.  That could be changed,
but it seems better to revise the definitions of gp node left/right/center
so that this phantom border is always 0 thickness.
Bug #2275