qt: fillbox would fail to draw any box for which y+height >= ymax
Because the y coordinate is inverted but height must be positive,
fillbox(x,y,width,height) is converted to fillbox(x,y+height,width,height).
But if y+height >= term->ymax then it is considered out of range and not
plotted at all.  Now we explicitly clip so that a box that reaches exactly
to the y border is not omitted from the plot.

Note that this was found because "clear" is implemented as a filled box,
and trying to clear a portion of the full screen would fail if
for origin + size == ymax.
Bug #2276