possible fix for dropped characters from piped command stream
There is a long history of bug reports where a single character is
dropped from the start of a command line following a successful plot.
Common features are piped input, mousing active, more likely to
happen if the plot window is closed interactively, use of gnu libreadline.
Not all of these are true for all reports.  To fix this we have at
various times added "ungetc('\n',stdin)" either in a specific terminal's
term->waitforinput() or most recently in mouse.c:event_reset().
The latest report (#2279) happens to be Windows-specific but otherwise
sounds just like the earlier ones.

Bugs #1293 #1559 #2134 #2279   and probably others

I have now found a possible culprit. At least it is a place where
the interactive and piped behavior differs.  Mouse interactions started
being multiplexed with readline input in 2000 (commit 2bf98f1b8).
In the original code and ever since has been the sequence
    if (term && term->waitforinput && interactive)
	return term->waitforinput();
    else
	return getc(fp);
This was later copied to parallel code for libedit and windows input.

But why should this be dependent on the state of "interactive"?
If the terminal is being moused then the program must deal with
events that come in via the mousing ipc, regardless of whether
stdin is piped or not.  So now I am removing the test on interactive.

I have not proved that this by itself fixes anything, since I do not
see the dropped character problem under linux from any gnuplot v5
verion.  If it *does* fix the problem genericly then the ungetc()
bandaid in mouse.c is no longer necessary.  I have marked it FIXME.