Allow build from current source not embedded in a git repository
Currently building the development version requires git to be installed
to fetch the `last modified' date from the commit information.  This is
okay when building from a git clone, but fails when downloading a snapshot
that does not contain the .git directory.  To this end, I implemented
a fallback which just uses the current date if git is not available.