qt: close window hotkey ('q' or ctrl-'q') should create only one event
Closing the plot window using 'q' or ctrl-q was generating two events,
one for the keypress itself and one for the close.  By itself this would
be OK but there was enough delay in processing the 'close' that another
plot may have intervened, causing it to close the wrong plot.
Now we suppress the keypress event and send only the 'close'
Bug #2098