New keyword "smooth kdensity period <val>"
For periodic data individual Gaussian components should be treated as repeating
at intervals of one period.  One example is data measured as a function of
angle, where the period is 2pi.  Another example is data indexed by day-of-year
and measured over multiple years, where the period is 365.
In such cases the period should be provided in the plot command:
     plot $ANGULAR_DAT smooth kdensity period 2*pi with lines

Feature Request #506