Interpret "set key at graph <x>, graph <y>" for 2D projections
In a 2D projection (set view projection {xz|yz}) the x and y axes
do not correspond to horizontal and vertical, so the normal
translation of the position specified by "set key at graph ..."
does not make sense.  Interpret the intent of the command as
"set key at graph <horizontal>, graph <vertical>".