reduce overhead of multi-parameter function calls
Internal routine f_calln() is invoked to handle user-defined functions
with multiple parameters. It was saving and restoring space for the
maximum allowable number of parameters on every invocation.  Now we
limit it to save/restore of the true number of parameters.   In the
case of vfill with a two parameter density function this yields a
~10% speed increase.