Deprecate the "boxwidth == -2" special case in "plot with boxerrorbars"
The boxerrorbars style had two different interpretations of 4 column input
normal:
      4 columns:  x  y  ydelta xdelta
special case if "set boxwidth -2":
      4 columns:  x  y  ylow  yhigh

Neither mode provided a way to simply use the current boxwidth
for data with explicit ylow/yhigh.

The special case is now deprecated in favor of the scheme
      3 columns:  x  y  ydelta
      4 columns:  x  y  ydelta xdelta         (xdelta <= 0 means use boxwidth)
      5 columns:  x  y  ylow  yhigh  xdelta   (xdelta <= 0 means use boxwidth)

The old special case is marked DEPRECATED in the user manual but
the code is still there so old scripts should continue to work.