Remove deprecated old-style call arguments $1 $2 etc
Support for this syntax was controlled by the configuration option
--enable-backwards-compatibility in version 5.2 and marked as deprecated
in the documentation.  In version 5.4 the configuration option was
removed, but unfortunately this had the opposite of the intended result
in that it turned support for the old syntax permanently on instead of
permanently off.  Now we remove the code altogether.
Bug #2298