Support "pointtype variable" with character point type for binary data files
Although there is currently no way to read string data directly from a binary
input file, it is nevertheless possible to have a using specifier that
evaluates to a string. Therefore we must be prepared to deal with various
possible uses of that string during binary input.  We already handled
"with labels" as a special case; now we also handle strings that are to
be used as a character point type for "with points pointtype variable".
TODO:  time strings are not handled
Bug #2300