New feature: local variables
The qualifier "local" in front of a variable assignment creates a new
variable of that name that shadows any previous variable with the same name.

The "local" qualifier is valid inside a "load" or "call" command,
or inside a clause delimited by curly brackets.  It is also valid in
function blocks (branch not yet committed).  It is ignored otherwise.
All of these contexts invoke lf_push() to save the previous program state.
When the corresponding lf_pop() is later executed the local variable goes
out of scope and the original variable, if any, is restored.

The original, shadowed variable is held in a temporary variable
GPLOCAL_xxx_NAME where xxx is the depth of push/pop stack and NAME is the
original variable name.