Minimal cleanup to interpretation of binary data
Revert 1be765b0 and 029ded4f in favor of a more general fix to the
problem of some plotting styles not having an entry in the table
of default using specs.  No default is needed if there is an actual
using spec.

TODO: get rid of the entire default table and most of the code that
      refers to it.  Is it so bad to require that a plot command
      reading binary data gives a 'using' spec?

Bug #2315