Minimal cleanup to interpretation of binary data
Revert 3186b737 and 3f5a74e8 in favor of a more general fix to the
problem of some plotting styles not having an entry in the table
of default using specs.  No default is needed if there is an actual
using spec.

TODO: Get rid of the entire default table and the code that uses it.
      Is it so bad to require that a plot command reading binary
      data gives a 'using' spec?

Bug #2315