emf terminal:  do not try to lseek a non-seekable output file
This was causing a double-free in the reported case, with the
complicating factor of being two levels deep in "load '-'".
That's a really weird state but I could not reproduce the fault without it.
This fixes the problem case, but may cause a spurious error message in
the case that someone really is trying to write an emf stream on stdout.

Bug #2327