handle leading space in dashtype specification
Up to now, leading space in a dashtype was totally ignored.
I.e. "dashtype '      .'  acted the same as dashtype '.'.
The effect of a blank dashtype was terminal dependent.
The qt terminal produced a solid line; the cairo terminals
reported a fatal error.

Now we always treat a blank dashtype as "solid".
If there are one or more leading space characters, they are
effectively replaced by adding a single space at the end.
I.e. dashtype "  -" acts as dashtype "- ".

Bug #2336