prevent invalid memory free operation on error exit from boxplot
The boxplot code temporarily replaced the coordinate array in the
plot structure and dummied it up to look like a candlestick plot.
But if an error exit occured while in this state freeing the
temporary coordinate array rather than the real array would be bad.
Close that possibility by doing a save-and-restore of the orignal
coordinate array content instead.