remove file national.h
The file national.h exists only to define the default values of
day/month names for strftime.  While it may be useful to have a default,
this function is now superseded by the system-supplied strftime in
conjunction with locale support.  Keep a minimal set of defaults
but move them into the locale handler in variable.c.