If event_reset is called from "set term", do not push character onto input stream
The code in event_reset contains a work-around for a bug that sometimes
loses one character from the terminal input stream if a CLOSE event is received.
The exact trigger conditions for this bug are not well understood.
The work-around is to push a newline character onto the head of the input
stream; if it is lost, no harm done, if it is not lost, the parser sees an
empty command line.  However in some circumstances (again not entirely understood)
the extra "\n" can itself cause a problem.  In particular if event_reset is
called from "set term" rather than from the mousing code, the extra "\n"
may do more harm than good.

TODO: perhaps "set term" does not really need to call event_reset?

Bug #2292 (see also Bug #2134)