jitter:  reassign slots in struct coordinate used to hold jitter offsets
The limited number of fields in struct coordinate means that data used
only by certain plot types needs to be stored in what would normally be
xlow, yhigh, etc.  This overloading is flagged by referring to the fields
by aliases CRD_XJITTER CRD_YJITTER and so on.  This makes it somewhat
easier to fix cases where overloading causes data loss.

The reassignment here prevents jitter from overwriting the point type field
CRD_PTTYPE in plots using "pointtype variable".
The cost is that "pointsize variable" data is overwritten instead.
That is less of a problem since the calculated jitter depends on
pointsize, so variable pointsize and jitter are already incompatible.

Bug #2354