new function round(x); range limits for int floor ceil round
Now that we have support for 64bit integers, it is a problem that
int(bignum) may be a representable integer but bignum itself has
insufficient precision to distinguish the difference between
bignum and bignum+1.

Test case:
   LEI = 9007199254740992  # Largest exact integer in IEEE754 double
   print int(LEI), int(LEI+1), int(LEI+2), int(LEI+3)
   print int(LEI-0.5), int(LEI+0.5)

Similar errors due to limited precision affect floor() ceil() round().

The problem is not simply the magnitude of the argument, it is the
number of significant digits.

int(x):  x > 2^63 returns NaN
         2^52 < x < 2^63 return trunc(x) even though int(real(x)) != x
floor(x):
ceil(x):
round(x): x > 2^52 return NaN