New terminal using block graphics
gnuplot terminal providing pseudo graphics using Unicode block characters
(halfs, quadrants, sextants), or Braille characters.  This is meant as an
alternative to the dumb, caca, or sixel terminals when a higher (pseudo)
resolution is required and sixel is not available.  It comes in handy
e.g. for remote use over browser based remote shells.  Reminiscent of
vintage computers like the ZX-81 and inspired by the drawille library.
The "best" character set to use (half, quad, sext, braille) depends on
the plot type.

Internally, graphics is handled by bitmap.c functions and the terminal is
hence only compiled in with the --with-bitmap-terminals configure option.
Text is dealt with by functions of the dumb terminal.  They support
enhanced text as well as UTF-8.  Text is always shown on top of the
graphics.

Color is supported, but limited to the actual character resolution, i.e.
in general the "pixels" in a charcell share the same color. If all pixels
are set, we change fore- and background colors to optimize this somewhat.
Otherwise, the average color of all pixel colors in a charcell is used.