webp terminal - single image or animation
This is a new terminal that shares routines from the pngcairo terminal
to produce images, then encodes the images using libwebp and libwebpmux.

Animations are created using the WebPAnimEncoder API.
They can be viewed using the vwebp utility or in most web browsers.

The webp format supports a huge number of encoding options but this
terminal uses only lossless 24-bit RGB with alpha channel.

TODO
- Bug: if the background is transparent, previous frames show through.
  There must be some API option to prevent this!
- Support creation of lossy animations?  This might require first
  implementing conversion to the YUV color space.