cairo terminals: Do not lose font info during multi-line enhanced text
Commit c95c679f was a work-around for a bug reported in 2015 that
bold/italic font properties got "stuck" after the key title is printed.
It introduced a reset to the default font at the end of enhanced text
processing.  However this caused another bug that the font was lost in
the middle of multi-line enhanced text.
Now we reset bold/italic flags but not the whole font.
It is still not clear to me why the cairo terminals need this yet
other terminals do not.
Bug #2378

TODO: The x11 terminal suffers from an equivalent bug but the analogous
change doesn't fix it.