Fix bookkeeping of input from files with multiple data sets
The index counter used during data input is incremented at the end of
every data set, before any data from the next set has been read.
This is fine while we are actually in the process of reading data but
leads to the index being one too high just after reading. So
  "plot FOO index n title sprintf("Index %g", column(-2))"
incorrectly places index n+1 in the title.
To fix this we introduce a new counter df_last_index_read that is not
updated until new data is actually read.
See discussion on Bug #2380

Also consolidate bookkeeping of pm3d 'at' options used for multiple
data sets with bookkeeping for all the other options.