If the plot title is a string constant, use it only once
Post-evaluation of plot titles in version 5.4 means that plotting from
a data file containing multiple data sets could generate a distinct
label for each one.  This may or may not be a good idea.
It is presumably more common to assign a single title to the entire data set.
In particular if the title is given as a string constant, why would we
want to repeat it? This patch treats that as a special case.

It leaves unresolved the question of whether to generate multiple
key entries for titles provided either as a string variable or as
a string expression.
Bug #2380