Allow 'splot with polygons' to use 'lc variable'
Variable color specified by 4th column of using specifier is taken from
the first vertex of the polygon.  For subsequent vertices of that same
polygon color information, if present at all, is ignored.

FIXME: This required a hack to datafile.c.  I would rather generalize
the "allow missing variable color column" treatment so that polygons are
are not a special case.