new routine: rgb_from_colorspec
Most of the graphics output code handles palette-based colors by
passing the palette index or gray to the terminal at the final stage
of output.  This doesn't work for newer 3D styles that use the pm3d
primitives because these require associating an RGB color with each
polygon at the time it is stored in the pm3d data structures, i.e.
prior to sorting, lighting, etc, and certainly prior to output.
The new routine rgb_from_colorspec() acts like the existing routine
apply_pm3dcolor() except that it returns the RGB value rather than
immediately sending it to the current terminal.

Currently used for 3D boxes and polygons.