bookkeeping for set/save/show pixmaps created from a colormap
For pixmaps read from a file we were already keeping the filename
so that it could be preserved across save/load.  Now if the pixmap
came from a colormap rather than a file we keep the colormap name.

Note: this is made trickier by the fact that we need to save
colormap data _before_ saving any pixmap that uses it. Otherwise
it will be read back in the wrong order during "load".
We do this by moving save_pixmaps out of save_set(0).
The alternative would be to save colormaps (and other arrays?)
before calling save_set rather than including them with other
variables dumped later.