pm: overhaul handling of font requests
Attempt to catch all variants of combinations of font requests with the
potential default font: empty, with or without font size or font name
and/or with or without italic/bold font attributes: 8 combinations each.
Remove all attempts to resolve some of these cases in the internal driver
code and let the external driver handle it.

We use strlcpy() and strlcat(), which are supported by klibC, to handle
string operations more safely. As EMX lacks these functions, we provide
our own simple approximations.

Extend the font aliases to include "sans", "serif", and "Times".