pm: font selection dialog
The font selection dialog was disabled in response to bug report #570. On
eComStation 1.2 (with some official fixes applied) the OS/2 bug is no
longer reproducable. This hence reintroduces this dialog, which now
supports bold and italic attributes and is always initialized from our
internal variables instead of relying on the data structure initialized
by a previous call.