Enable compilaton with gcc 9 and kLIBC on OS/2
This enables using a newer compiler by at least 6 major releases. Many
libraries are available for it using rpm repositories. This enables
compilation of the cairo terminals (including webp), the gd terminals,
and the lua (tikz) terminal. Also, Qt5 becomes available and the qt
terminal might hence be enabled later.

Despite the significantly evolved C library, kLIBC, we only have to
address two significant issues.  With the modern kLIBC, spawn() and
friends cannot execute a GUI program from a console programme. So we use
the command shell to start a new driver in the background. For command
line input, getc(stdin) does not seem to work, so we use conio.h's
getch() instead as a work-around.