pm: intermediate functions to communicate with the outboard driver
This makes the code more readable and less error-prone. It also reveals some
inconsistencies in the protocol. But since the gnupmdrv external driver is
also used by third-parties, it is too late to change them now. Padding of
strings with NULs did not match the expectation on the external driver side.