pm: transparency mask for RGBA images
The OS/2 GPI does not support alpha level transparency, but this patch
supports a transparency mask (bilevel). The method requires drawing the
mask and the image data separately. This works as long as transparent
pixels ar all black. The code contains also a more accurate version which
requires a third drawing operation.
Since pixmaps are always drawn as RGBA, we optimize the case when the
image does not really contain transparency and send an RGB image.