Do not draw user-defined tics that lie outside a nonlinear axis range
For linear axes there is a "slop" region at each end of the axis range
such that if you request a tic mark slightly outside the range it will
be drawn as requested.
Historically this slop has been set to axis_range / 1000.
For log-scale and other nonlinear axes this makes little sense,
so we now set it to zero.

Bug #2408