Say goodbye to EMX on DOS and the emxvga terminal
This removes support for EMX on DOS. The EMX/GCC compiler was last updated
about 20 years ago with its gcc is at version 2.8.1.  We also remove the
emxvga terminal which supported DOS and OS/2 fullscreen sessions. This
terminal offered only very basic features. On DOS, the pc (OpenWatcom) and
svga (DJGPP) terminals are much better options. [An attempt to somewhat
modernize this terminal can be found in the "emxvga-terminal" branch of my
git fork on SF.]
Note that the remaining preprocessor tests for __EMX__ are still required
on OS/2.