pm: display status text in status bar
This adds a status bar to the bottom of the pm terminal window to hold the
mouse status information, much like the wxt, qt, and win terminals.
The status text cannot be updated in the pipe reader thread directly, so it
is simply postponed to the next message queue event.