New: uigamma(a,x) provides regularized upper incomplete gamma function
uigamma(a,x) = Q(a,x)
igamma(a,x)  = P(a,x)
P(a,x) + Q(a,x) = 1.0

The motivation is that approximating Q as (1-P) loses precision
as P approaches 1 (see uigamma.dem).

This implementation is from the Cephes library (Moshier 2000).
The domain is limited to positive real a, x.

TODO: replace or extend to handle complex arguments