Add uigamma as a sample plugin usable in earlier gnuplot versions
The plugin demo uses only toy functions.
This adds an example that people might actually want to use,
the regularized upper incomplete gamma function
[Abramowitz & Stegun Q(a,x) NIST DLMF 8.2.4].

It contains the same code as in version 5.5 for uigamma(a,x),
but since it is a plugin it is also usable by any version of
gnuplot that supports plugins.