Facilitate and document use of a plugin to provide uigamma function
Some users may want a particular function that is not present in
their current version of gnuplot, whether because it isn't in the
distribution at all or because whoever provided their gnuplot binary
did not configure it in.

The new section in the user manual on plugins, together with new files
demo/plugin/README and uigamma_plugin.c, show how you can use the
plugin mechanism to add a specific function without having to edit,
reconfigure, and rebuild your gnuplot executable.  It uses the
upper incomplete gamma function uigamma(a,x) as an example.

uigamma is provided as a built-in function in the gnuplot development
version but not in earlier versions, so there is a real use case for
this plugin over and above serving as an example.