memory leak if plot iterator is not freed before reuse
The allocate/free is supposedly done in plot and splot, but long runs
with valgrind found some that were never freed.  I am not sure what
exactly the case is that causes this, but now we make sure the previous
plot iterator was freed before starting a new plot/splot.