wxt: discard events that arrive prior to terminal initialization
The sequence below could cause memory corruption if an event arrived
between the second "set term" command and display of the subsequent plot.
If the "replot on resize" widget was enabled, this was likely to happen.
	set term wxt 1
	plot foo
	set term wxt 2
	plot baz

I only saw this with mono-threaded wxt, but I worry the multi-threaded
wxt version could hit it also.