wxt: remove monothreaded interlock kludge
wxt_interlock was added in 2018 commit 17c0d770 as a kludge to prevent
recursion by a call call to open a new window after "set term wxt N".
This seems to be no longer needed, either because of a change in wxgtk
or possibly because of the recent fix in commit 3934fbc5.
I can't confirm either hypothesis but also I can no longer trigger the failure.
This reverts 17c0d770.