qt terminal: sanity checks when leaving gnuplot core to enter "persist" mode
If there are no visible plot windows for gnuplot_qt to manage,
exit immediately.  This is needed because window manager actions that
"close" a window may only hide it.  If it is still there invisibly then
gnuplot_qt will hang around as a zombie waiting for interaction with an
invisible window that cannot happen.

Bug #2188