Plug memory leak in parse_array_assignment_expression()
Any expression that starts with an array reference is first parsed
as if it were an array element assignment.  If this doesn't work,
it is parsed again.  Unfortunately there was incomplete cleanup of the
first, failed, parsing attempt.
    array A = [1,2,3,4]
    print A[strlen("XX")]
    print A[ sum[i=1:2]i ]
leaked the memory allocated for PUSHC "XX" and for the action table
stored to evaluate the summation.