Revise auto-placement of tics for time data
For timescales of days or longer, if a ticmark is generated for
some date then place it at the start of that same day (0 AM).
The spacing between "day" tics is always exactly N days.

It used to be that the date was rounded up if time > 22:00 but
hours/minutes/seconds were left unchanged. This meant that the
tic positions were not exactly on a day boundary and might not
even be on the same day. Thus the spacing between "dates" was not
an integral number of days.

This discrepancy was particularly obvious when using the new
option "set minitics time N days", since the evenly spaced
minitics did not align with the major tics.