New: "sparse matrix" input mode for incomplete grid data
    plot FOO sparse matrix=(ncols,nrows) [origin=(x0,y0)] [dx=<val>] [dy=<val>]

The immediate motivation is to allow "plot with image" to handle pixels
that are out of order or missing altogether, for example a heat map where
some conditions were not measured.
Example:
    plot '-' sparse matrix=(3,3) origin=(1,1) with image
    1 1 5
    3 3 2
    2 2 10
    e
This example creates a full 3x3 pixel array with all the off-diagonal pixels
set to UNDEFINED.

Each pixel of the grid defined in the plot command is initialized to type
UNDEFINED and the x/y coordinates are initialized to match the definition.

Individual pixel values [x y z] are then read from the input file.
If the x y coordinates match the grid definition the result is a full matrix,
possibly out of order.  Otherwise some of the grid pixels remain UNDEFINED.
Input coordinates that lie outside the defined grid generate a warning but
are otherwise ignored.