Fix handling Window messages for input from a pipe
As pointed out in bug report #2412, MsgWaitForMultipleObjects() does not work
as expected if the input comes from a pipe. This reverts 651af6267b
"Wait for pipe events without additional thread", which aimed at speeding-up
piped input, Instead of recreating the input thread for every single
character, we now keep the thread alive and signal an event object on new
input. This should still be much faster.

Bugs #2412 and #2204