dumb: wide character support
Wide characters occupy two character cells and hence currently distort
the graph layout. We now mark the subsequent cell to a wide char as
"occupied" with a NUL. These cells are not printed on output, thus
restoring alignment.

Calls to set_pixel() blank the cell to the left, if the current cell
is marked as occupied, and the one to the right, if they overwrite a wide
character.

This patch does not yet handle overprinting of text made of wide
characters. Also, the logic to test for wide characters is incomplete but
equivalent to mbwidth() in readline.c and strwidth_utf8() in estimate.trm.

These changes also apply to the "block" terminal as the text routines are
shared.