better management of temporary arrays by CALL, CALLN, and POP
Built-in operators that take an array parameter, e.g. |A| or A[m:n],
free the memory allocated for A if it is marked TEMP_ARRAY.
In order for user-defined function to present the same behavior they
must (1) check and free any TEMP_ARRAY passed to them and
(2) make sure that the same array is not freed underneath by a call
to some built-in function or operator.  To this end we now switch
the TEMP_ARRAY flag to ARRAY on entry to f_call or f_calln, and
free the array on exit.

Similarly, the values resulting from the left-most expressions in a
comma-separated serial evaluation are discarded.  If this value is
a TEMP_ARRAY, free it.

It is still possible to leak memory if an error trap occurs during
evaluation of an expression that creates a TEMP_ARRAY.

Supersedes previous incomplete treatment added by commit 6441c36d