initialize value to NOTDEFINED so that failed evaluation leaves it that way
Fuzzing found that evaluation of a malformed "if" condition could leave
the resulting supposedly 0/1 value undefined instead. Although we could
trap this specific case, it seems possible that the same thing could bite
whenever the result of a mal-formed evaluation is used.
Now we initialize the return value to NOTDEFINED, so that if the evaluation
fails to do anyhing the "result" is at least is in a known state and testable.